+++ MSAL/test/integration/native_auth/end_to_end/mfa/MSALNativeAuthSignInJITEndToEndTests.swift
@@ -52,13 +52,13 @@ final class MSALNativeAuthSignInJITEndToEndTests: MSALNativeAuthEndToEndPassword
 
         guard signInDelegateSpy.onSignInStrongAuthMethodRegistrationCalled,
               let strongAuthState = signInDelegateSpy.newStateStrongAuthMethodRegistration,
-              let authMethod = signInDelegateSpy.authMethods?.first else {
+              let authMethod = signInDelegateSpy.authMethods?.first(where: { $0.channelTargetType.isEmailType }) else {
             XCTFail("Sign in failed or strong auth method registration not required")
             return
         }
 
-        // Step 3: Add Strong Auth Method, but don't specify verification contact so it's preverified
-        let challengeParameters = MSALNativeAuthChallengeAuthMethodParameters(authMethod: authMethod)
+        // Step 3: Add Strong Auth Method
+        let challengeParameters = MSALNativeAuthChallengeAuthMethodParameters(authMethod: authMethod, verificationContact: username)
         let challengeExpectation = expectation(description: "challenging auth method")
         let challengeDelegateSpy = RegisterStrongAuthChallengeDelegateSpy(expectation: challengeExpectation)
 
@@ -97,15 +97,14 @@ final class MSALNativeAuthSignInJITEndToEndTests: MSALNativeAuthEndToEndPassword
 
         guard signInDelegateSpy.onSignInStrongAuthMethodRegistrationCalled,
               let strongAuthState = signInDelegateSpy.newStateStrongAuthMethodRegistration,
-              let authMethod = signInDelegateSpy.authMethods?.first else {
+              let authMethod = signInDelegateSpy.authMethods?.first(where: { $0.channelTargetType.isEmailType }) else {
             XCTFail("Sign in failed or strong auth method registration not required")
             return
         }
 
         // Step 3: Add Strong Auth Method and specify different email
         let newEmail = generateSignUpRandomEmail()
-        let challengeParameters = MSALNativeAuthChallengeAuthMethodParameters(authMethod: authMethod)
-        challengeParameters.verificationContact = newEmail
+        let challengeParameters = MSALNativeAuthChallengeAuthMethodParameters(authMethod: authMethod, verificationContact: newEmail)
         let challengeExpectation = expectation(description: "challenging auth method")
         let challengeDelegateSpy = RegisterStrongAuthChallengeDelegateSpy(expectation: challengeExpectation)
 
@@ -168,15 +167,14 @@ final class MSALNativeAuthSignInJITEndToEndTests: MSALNativeAuthEndToEndPassword
 
         guard signInDelegateSpy.onSignInStrongAuthMethodRegistrationCalled,
               let strongAuthState = signInDelegateSpy.newStateStrongAuthMethodRegistration,
-              let authMethod = signInDelegateSpy.authMethods?.first else {
+              let authMethod = signInDelegateSpy.authMethods?.first(where: { $0.channelTargetType.isEmailType }) else {
             XCTFail("Sign in failed or strong auth method registration not required")
             return
         }
 
         // Step 3: Add Strong Auth Method and specify different email
         let newEmail = generateSignUpRandomEmail()
-        let challengeParameters = MSALNativeAuthChallengeAuthMethodParameters(authMethod: authMethod)
-        challengeParameters.verificationContact = newEmail
+        let challengeParameters = MSALNativeAuthChallengeAuthMethodParameters(authMethod: authMethod, verificationContact: newEmail)
         let challengeExpectation = expectation(description: "challenging auth method")
         let challengeDelegateSpy = RegisterStrongAuthChallengeDelegateSpy(expectation: challengeExpectation)
 
@@ -239,7 +237,7 @@ final class MSALNativeAuthSignInJITEndToEndTests: MSALNativeAuthEndToEndPassword
 
         guard signInDelegateSpy.onSignInStrongAuthMethodRegistrationCalled,
               let strongAuthState = signInDelegateSpy.newStateStrongAuthMethodRegistration,
-              let authMethod = signInDelegateSpy.authMethods?.first else {
+              let authMethod = signInDelegateSpy.authMethods?.first(where: { $0.channelTargetType.isEmailType }) else {
             XCTFail("Sign in failed or strong auth method registration not required")
             return
         }
@@ -248,6 +246,49 @@ final class MSALNativeAuthSignInJITEndToEndTests: MSALNativeAuthEndToEndPassword
         XCTAssertTrue(signInDelegateSpy.onSignInPasswordErrorCalled)
         XCTAssertTrue(signInDelegateSpy.error?.isBrowserRequired ?? false)
         XCTAssertNotNil(signInDelegateSpy.error?.errorDescription)
+    }
+    
+    func test_createUserAndAddInvalidPhoneAsStrongAuthMethod_thenInvalidInputErrorIsReturned() async throws {
+        throw XCTSkip("SMS auth method not yet available in lab tenant")
+#if os(macOS)
+        throw XCTSkip("For some reason this test now requires Keychain access, reason needs to be investigated")
+#endif
+        let username = generateSignUpRandomEmail()
+        // Step 1: Create User
+        guard let signInAfterSignUpState = await signUpInternally(username: username, password: generateRandomPassword(), application: initialisePublicClientApplication()) else {
+            XCTFail("onSignUpCompleted not called or state is nil")
+            return
+        }
+
+        // Step 2: Attempt to Sign In after signUp
+        let signInExpectation = expectation(description: "signing in")
+        let signInDelegateSpy = SignInAfterSignUpDelegateSpy(expectation: signInExpectation)
+
+        let signInParameters = MSALNativeAuthSignInAfterSignUpParameters()
+        signInParameters.claimsRequest = MSALClaimsRequest(jsonString: "{\"access_token\":{\"acrs\":{\"essential\":true,\"value\":\"c4\"}}}", error: nil)
+        signInAfterSignUpState.signIn(parameters: signInParameters, delegate: signInDelegateSpy)
+
+        await fulfillment(of: [signInExpectation])
+
+        guard signInDelegateSpy.onSignInStrongAuthMethodRegistrationCalled,
+              let strongAuthState = signInDelegateSpy.newStateStrongAuthMethodRegistration,
+              let smsAuthMethod = signInDelegateSpy.authMethods?.first(where: { $0.channelTargetType.isSMSType }) else {
+            XCTFail("Sign in failed or strong auth method registration not required or SMS method not found")
+            return
+        }
+
+        // Step 3: Add Strong Auth Method and specify invalid phone number
+        let invalidPhone = "+123" // Clearly invalid phone number
+        let challengeParameters = MSALNativeAuthChallengeAuthMethodParameters(authMethod: smsAuthMethod, verificationContact:  invalidPhone)
+        let challengeExpectation = expectation(description: "challenging auth method with invalid phone")
+        let challengeDelegateSpy = RegisterStrongAuthChallengeDelegateSpy(expectation: challengeExpectation)
+
+        strongAuthState.challengeAuthMethod(parameters: challengeParameters, delegate: challengeDelegateSpy)
+
+        await fulfillment(of: [challengeExpectation])
+
+        XCTAssertTrue(challengeDelegateSpy.onRegisterStrongAuthChallengeErrorCalled)
+        XCTAssertEqual(challengeDelegateSpy.error?.isInvalidInput, true)
     }
     
     // MARK: private methods