+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [2.5.0]
+* Awaiting MFA Delegate now automatically returns the AuthMethods to be used when calling MFA Challenge #2687
+* SDK now handles SMS as strong authentication method #2722
+
 ## [2.4.2]
 * Bug fix to prevent crash when deallocating STK SecKeyRef